script/requirements.txt
numpy
pandas
scikit-learn
pinecone-client
python-dotenv